Move to using a dictionary to provide options for choice windows in javascript
